fix: factor out shared tasks (#529)
